+++ lib/frames/frame-symbol.ts
@@ -116,6 +116,13 @@ export class FrameOperator extends FrameSymbol {
   public static override readonly SIGIL_STARTS: readonly SigilStart[] = [
     { key: FrameOperator.OPERATOR_START.toString(), mode: "atom" },
   ];
+  public readonly operator: string;
+
+  constructor(source: string, meta: Context = NilContext) {
+    super(source, meta);
+    this.operator = source;
+    this.is.operator = true;
+  }
 
   public static operator_chars(): string {
     return "&|?:+\\-*%<>!";
@@ -127,7 +134,11 @@ export class FrameOperator extends FrameSymbol {
   }
 
   public override in(_contexts: Frame[] = [Frame.nil]): Frame {
-    return FrameSymbol.for(this.data);
+    return this;
+  }
+
+  public override called_by(context: Frame): Frame {
+    return FrameSymbol.for(this.data).called_by(context);
   }
 
   public override string_start(): string {